refactor(noncomp): reuse the SVM xoshiro RNG; self-contained comments
Address two review comments on the history sampler.

Switch the sampler from std::mt19937_64 to Xoshiro256PlusPlus, the same
scientific PRNG the SVM uses, hoisted out of svm.h into a shared
util/xoshiro.h that both now include. The sampler keeps its
cross-platform fixed-seed reproducibility from a single generator
instead of a second one; the [0, 1) extraction is unchanged.

Drop the design-doc section/step references ("section 4.2", "step 3",
"the MVP") from the sampler, stepper, and model comments so they stand
on their own, since the design note is meant to be removed eventually.
No behavior change; full Debug suite green.
